remove old serieschart and rename backtestchart
