+++ frontend/src/components/BacktestChart.tsx
